Fix reading mbox: always start parser from 'init state except for first mail
The parser for reading mail from stream is called both for the rebuild stage
(when token tables are eventually established) and the classification stage
in which only a single mail is most commonly parsed (but this is not required).

The first mail may therefore come from either the top of a mbox file (where mail
entries are preceded with From lines) or from a pipe (stdin) being fed by
procmail or what have you. Because of this, we need to be flexible and start
parsing the first mail either with or without seeing a From line first.

In case a From line is required, the parser will be more robust, because any
garbage (or skipped attachment from an earlier mail) will be discarded and not
considered part of any mail header or body. This is normally what we want, and
is now the case for all mail except - for reasons explained above - the first
one in the stream.