Fix tokenizer for the case of '<' not actually starting a HTML comment
The tokenizer takes special care that HTML comments are discarded in a
way that they don't even constitute token separators. For this reason,
a whole mini-state-machine is implemented in the tokenizer function.

When the last character read is a '<' (potentially starting a HTML
comment with '<!--'), the state is transitioned to 'tokstate-html-in1.
When, after being in this state, the next character is *not* '!', we
know it is not a HTML comment. So that character must be saved to the
token buffer. However, since '<' is not a token constituent, the
previously accumulated token buffer must be emitted, breaking the
token at the point of '<' and thus acting the same way as it would
without the added complexity of discarding HTML comments.