Fix decode-html-entities logic
The logic of finding &entity; substrings was buggy.

It may happen that a string contains multiple '&' characters
before the ';', especially with HTML links. In such case, the
last '&' before the ';' must be used. This is now fixed by
searching backwards from ';' for the '&'.

An example link that triggered the buggy behavior:
<a href="http://[...]/mail/unsubscribe.php?M=4325143&C=ac84a87a&L=70&N=69">Leiratkoz&aacute;s</a>